+++ src/modules/calendar/v1/interfaces/calendar-service.interface.ts
@@ -1,6 +1,6 @@
-import { CreateCalendarRequest } from 'src/dto/calendar-event/requests/create-calendar.request';
-import { UpdateCalendarRequest } from 'src/dto/calendar-event/requests/update-calendar.request';
-import { CalendarListResponse } from 'src/dto/calendar-event/responses/calendar-list.response';
+import { CreateCalendarRequest } from 'src/dto/calendar/requests/create-calendar.request';
+import { UpdateCalendarRequest } from 'src/dto/calendar/requests/update-calendar.request';
+import { CalendarListResponse } from 'src/dto/calendar/responses/calendar-list.response';
 import { PycUser } from 'src/dto/common/dto/pyc-user.dto';
 
 export interface ICalendarService {
@@ -10,7 +10,7 @@ export interface ICalendarService {
   getCalendarsByMonth(
     pycUser: PycUser,
     monthDate: Date,
-    options?: { offset: number; limit: number },
+    options?: { offset?: number; limit?: number },
   ): Promise<CalendarListResponse>;
   //U
   update(pycUser: PycUser, targetId: number, req: UpdateCalendarRequest): Promise<void>;
