+++ src/modules/calendar/v1/services/calendar.service.ts
@@ -1,8 +1,8 @@
 import { Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CreateCalendarRequest } from 'src/dto/calendar-event/requests/create-calendar.request';
-import { UpdateCalendarRequest } from 'src/dto/calendar-event/requests/update-calendar.request';
-import { CalendarListResponse } from 'src/dto/calendar-event/responses/calendar-list.response';
+import { CreateCalendarRequest } from 'src/dto/calendar/requests/create-calendar.request';
+import { UpdateCalendarRequest } from 'src/dto/calendar/requests/update-calendar.request';
+import { CalendarListResponse } from 'src/dto/calendar/responses/calendar-list.response';
 import { PycUser } from 'src/dto/common/dto/pyc-user.dto';
 import { Calendar } from 'src/entities/calendar-event/calendar.entity';
 import { Church } from 'src/entities/church/church.entity';
